+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Catalyst SD-WAN Lab 2.1.5 [unreleased]
 
+- In restore task, fix issue where task might fail if Manager and Controller/Validator versions are not matching
 - In restore task, fix issue where the final output contains port 8443 instead of port used by PATty
 
 # Catalyst SD-WAN Lab 2.1.4 [Sep 10, 2025]